requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.38.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.18.0
numpy==1.26.4
openai==2.26.0
packaging==23.2
pandas==2.3.3
Pillow==10.0.0
protobuf==4.25.8
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
referencing==0.37.0
requests==2.31.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
